feat: 新增 popup 组件
- 新增 popup 附着弹出层组件，将作为「附着弹出、选择器」等类型组件的基础组件
- 迁移 dropdown 为继承 popup 的实现方式
- 迁移 colorpicker 为继承 popup 的实现方式
- 新增 menu 组件，从 dropdown 中解耦
- 优化 `lay.position` 部分选项命名
- 微调 `lay-upbit` 动画细节
