+++ src/css/modules/colorpicker.css
@@ -1,5 +1,7 @@
-/** 颜色选择器 **/
-.lay-colorpicker {
+/** colorpicker.css **/
+
+/* 颜色选择器盒子 */
+.lay-colorpicker-box {
   width: 38px;
   height: 38px;
   border: 1px solid var(--lay-border-color);
@@ -12,21 +14,21 @@
   transition: all 0.3s;
   box-sizing: border-box;
 }
-.lay-colorpicker:hover {
+.lay-colorpicker-box:hover {
   border-color: var(--lay-gray-400);
 }
-.lay-colorpicker.lay-colorpicker-lg {
+.lay-colorpicker-box.lay-colorpicker-lg {
   width: 44px;
   height: 44px;
   line-height: 30px;
 }
-.lay-colorpicker.lay-colorpicker-sm {
+.lay-colorpicker-box.lay-colorpicker-sm {
   width: 30px;
   height: 30px;
   line-height: 20px;
   padding: 3px;
 }
-.lay-colorpicker.lay-colorpicker-xs {
+.lay-colorpicker-box.lay-colorpicker-xs {
   width: 22px;
   height: 22px;
   line-height: 16px;
@@ -56,17 +58,7 @@
 }
 
 .lay-colorpicker-main {
-  position: absolute;
-  left: -999999px;
-  top: -999999px;
-  z-index: 77777777;
   width: 280px;
-  margin: 5px 0;
-  padding: 7px;
-  background: #fff;
-  border: 1px solid var(--lay-border-color);
-  border-radius: var(--lay-border-radius);
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.12);
 }
 .lay-colorpicker-main-wrapper {
   height: 180px;
