+++ src/css/modules/popup.css
@@ -1,0 +1,38 @@
+/* popup.css */
+
+.lay-popup {
+  --lay-popup-gap: 5px;
+
+  position: absolute;
+  left: -999999px;
+  top: -999999px;
+  z-index: 77777777;
+  margin: var(--lay-popup-gap) 0;
+  box-sizing: border-box;
+
+  * {
+    box-sizing: border-box;
+  }
+}
+.lay-popup::before {
+  content: '';
+  position: absolute;
+  width: 100%;
+  height: var(--lay-popup-gap);
+  left: 0;
+  top: calc(-1 * var(--lay-popup-gap));
+}
+.lay-popup-content {
+  position: relative;
+  padding: var(--lay-spacing-sm);
+}
+.lay-popup-backdrop {
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  position: fixed;
+  pointer-events: auto;
+  background-color: rgb(0, 0, 0);
+  opacity: 0.03;
+}
