+++ src/index.umd.js
@@ -16,11 +16,12 @@ import * as utils from './utils/index.js';
 
 // 导入组件模块
 import { layer } from './components/layer.js';
-import { laydate } from './components/laydate.js';
+import { popup } from './components/popup.js';
 import { laypage } from './components/laypage.js';
+import { laydate } from './components/laydate.js';
 import { dropdown } from './components/dropdown.js';
-import { slider } from './components/slider.js';
 import { colorpicker } from './components/colorpicker.js';
+import { slider } from './components/slider.js';
 import { nav } from './components/nav.js';
 import { breadcrumb } from './components/breadcrumb.js';
 import { collapse } from './components/collapse.js';
@@ -51,11 +52,12 @@ const layui = {
   Component,
   utils,
   layer,
-  laydate,
+  popup,
   laypage,
+  laydate,
   dropdown,
-  slider,
   colorpicker,
+  slider,
   nav,
   breadcrumb,
   collapse,
