+++ tests/visual/popup.html
@@ -1,0 +1,233 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <meta
+      name="viewport"
+      content="width=device-width, initial-scale=1, maximum-scale=1"
+    />
+    <title>popup 附着弹出层 - Layui</title>
+    <link rel="stylesheet" href="./assets/dist/css/layui.css" />
+  </head>
+  <body>
+    <div class="lay-container lay-padding-3">
+      <h2>常规</h2>
+      <hr />
+      <div
+        style="display: flex; align-items: center; gap: 6px; flex-wrap: wrap"
+      >
+        <button class="lay-btn" id="demo-normal">普通示例</button>
+        <button class="lay-btn" id="demo-backdrop">开启遮罩</button>
+      </div>
+
+      <br /><br />
+      <h2>触发方式</h2>
+      <hr />
+      <div
+        style="display: flex; align-items: center; gap: 6px; flex-wrap: wrap"
+      >
+        <button class="lay-btn" id="demo-hover">hover</button>
+        <button class="lay-btn" id="demo-mousedown">mousedown</button>
+        <button class="lay-btn" id="demo-dblclick">dblclick</button>
+      </div>
+
+      <br /><br />
+      <h2>打开动画</h2>
+      <hr />
+      <div
+        style="display: flex; align-items: center; gap: 6px; flex-wrap: wrap"
+      >
+        <button class="lay-btn" id="demo-downbit">downbit （默认）</button>
+        <button class="lay-btn" id="demo-upbit">upbit</button>
+        <button class="lay-btn" id="demo-scale">scale</button>
+        <button class="lay-btn" id="demo-scaleSpring">scaleSpring</button>
+        <button class="lay-btn" id="demo-scalesmall">scalesmall</button>
+        <button class="lay-btn" id="demo-fadein">fadein</button>
+      </div>
+
+      <br /><br />
+
+      <h2>设置样式</h2>
+      <hr />
+      <div
+        style="display: flex; align-items: center; gap: 6px; flex-wrap: wrap"
+      >
+        <button class="lay-btn" id="demo-width">控制最小最大宽度</button>
+        <button class="lay-btn" id="demo-style">自定义任意样式</button>
+        <button class="lay-btn" id="demo-content">自定义样式与内容</button>
+      </div>
+
+      <br /><br />
+
+      <h2>外部操作</h2>
+      <hr />
+      <div
+        style="display: flex; align-items: center; gap: 6px; flex-wrap: wrap"
+      >
+        <div id="demo-external">
+          <strong>目标元素</strong>
+          （通过右侧按钮打开或关闭面板）
+        </div>
+        <div>
+          <button class="lay-btn" lay-on="open">open</button>
+          <button class="lay-btn" lay-on="close">close</button>
+        </div>
+      </div>
+      <br />
+      <div
+        style="display: flex; align-items: center; gap: 11px; flex-wrap: wrap"
+      >
+        <label for="updateContent">模拟 autocomplete 效果</label>
+        <div>
+          <input
+            type="text"
+            id="updateContent"
+            placeholder="请输入"
+            class="lay-input"
+          />
+        </div>
+      </div>
+
+      <br /><br />
+    </div>
+
+    <script src="./assets/dist/layui.js"></script>
+    <script>
+      layui.use(() => {
+        const { popup, tabs, lay, $, utils } = layui;
+
+        // 常规
+        popup.render({
+          elem: '#demo-normal',
+          content: '一个普通的附着式弹出层',
+        });
+
+        // 开启遮罩
+        popup.render({
+          elem: '#demo-backdrop',
+          content:
+            'Popup 是附着弹出、选择器等类型组件（如：dropdown、laydate、colorPicker）的基础组件。',
+          style: 'width: 350px;',
+          backdrop: 0.5,
+        });
+
+        // 触发方式
+        ['hover', 'mousedown', 'dblclick'].forEach((trigger) => {
+          popup.render({
+            elem: `#demo-${trigger}`,
+            trigger,
+            content: `通过 ${trigger} 事件打开`,
+          });
+        });
+
+        // 打开动画
+        [
+          'downbit',
+          'upbit',
+          'scale',
+          'scaleSpring',
+          'scalesmall',
+          'fadein',
+        ].forEach((anim) => {
+          popup.render({
+            elem: `#demo-${anim}`,
+            content: `${anim} 动画方式弹出`,
+            anim,
+          });
+        });
+
+        // 控制最小最大宽度
+        popup.render({
+          elem: '#demo-width',
+          content: '通过 style 选项设置最小最大宽度',
+          style: 'min-width: 300px; max-width: 600px;',
+        });
+        // 自定义任意样式
+        popup.render({
+          elem: '#demo-style',
+          content: '通过 style 选项设置任意样式',
+          style:
+            'min-width: 300px; max-width: 600px; background-color: #16b777; color: #fff; border-radius: 11px; padding: 11px;',
+        });
+        // 自定义样式与内容
+        popup.render({
+          elem: '#demo-content',
+          content: `
+<div class="lay-tabs">
+  <ul class="lay-tabs-header">
+    <li class="layui-this">Tab Header1</li>
+    <li>Tab Header2</li>
+    <li>Tab Header3</li>
+  </ul>
+  <div class="lay-tabs-body">
+    <div class="lay-tabs-item layui-text layui-show"><p style="padding-bottom: 10px;">在 content 属性中传入任意的 html 内容，可替代默认的下拉菜单结构，从而实现更多有趣的弹出内容。</p><p> 是否发现，dropdown 组件不仅仅只是一个下拉菜单或者右键菜单，它能被赋予许多的想象可能。</p></div>
+    <div class="lay-tabs-item">Tab Content2</div>
+    <div class="lay-tabs-item">Tab Content3</div>
+  </div>
+</div>
+          `,
+          className: 'demo-dropdown-tabs',
+          style: 'width: 390px; box-shadow: 1px 1px 30px rgb(0 0 0 / 12%);',
+          afterOpen(options, $rootElem) {
+            tabs.render({
+              elem: $rootElem.find('.lay-tabs'),
+            });
+          },
+        });
+
+        // 外部打开或关闭面板
+        popup.render({
+          elem: '#demo-external',
+          trigger: 'manual',
+          content: '弹出层内容',
+        });
+
+        utils.on(
+          {
+            open: () => {
+              popup.open('demo-external');
+            },
+            close: () => {
+              popup.close('demo-external');
+            },
+          },
+          { trigger: 'mouseenter' },
+        );
+
+        // 模拟 autocomplete
+        const autocompletInst = popup.render({
+          elem: '#updateContent',
+          trigger: 'focus',
+          content: '初始内容',
+          style: 'min-width: 168px;',
+        });
+        const autocompletInstId = autocompletInst.options.id;
+        const sourceContent = autocompletInst.options.content;
+
+        // 输入框输入事件
+        autocompletInst.options.$elem.on('input', function () {
+          const $this = $(this);
+          const value = $this.val().trim();
+
+          // 输入为空时，恢复初始渲染状态
+          if (!value) {
+            popup.updateContent(autocompletInstId, sourceContent);
+            return;
+          }
+
+          // 模拟生成随机行内容
+          const randomContent = Array(Math.floor(Math.random() * 5) + 1)
+            .fill(0)
+            .map(
+              (_, i) =>
+                `${i + 1}. <strong>${lay.escape(value)}</strong>${Math.random().toString(32).substring(2)}`,
+            )
+            .join('<br />');
+
+          // 更新弹出层内容
+          popup.updateContent(autocompletInstId, randomContent);
+        });
+      });
+    </script>
+  </body>
+</html>
